Rewrote list comparison unit tests
